+++ python/cuml/test/test_metrics.py
@@ -230,28 +230,24 @@ def test_rand_index_score(name, nrows):
 @pytest.mark.parametrize('metric', (
     'cityblock', 'cosine', 'euclidean', 'l1', 'sqeuclidean'
 ))
-def test_silhouette_score(metric, labeled_clusters):
-    X, labels = labeled_clusters
-    cuml_score = cu_silhouette_score(X, labels, metric=metric)
-    sk_score = sk_silhouette_score(X, labels, metric=metric)
-    assert_almost_equal(cuml_score, sk_score)
-
-@pytest.mark.parametrize('metric', (
-    'cityblock', 'cosine', 'euclidean', 'l1', 'sqeuclidean'
-))
-def test_silhouette_score_batched(metric, labeled_clusters):
+@pytest.mark.parametrize('chunk_divider', [1, 3, 5])
+def test_silhouette_score_batched(metric, chunk_divider, labeled_clusters):
     X, labels = labeled_clusters
     cuml_score = cu_silhouette_score(X, labels, metric=metric,
-                                     chunksize=X.shape[0] / 10)
+                                     chunksize=int(X.shape[0]/chunk_divider))
     sk_score = sk_silhouette_score(X, labels, metric=metric)
     assert_almost_equal(cuml_score, sk_score)
+
 
 @pytest.mark.parametrize('metric', (
     'cityblock', 'cosine', 'euclidean', 'l1', 'sqeuclidean'
 ))
-def test_silhouette_samples(metric, labeled_clusters):
+@pytest.mark.parametrize('chunk_divider', [1, 3, 5])
+def test_silhouette_samples_batched(metric, chunk_divider, labeled_clusters):
     X, labels = labeled_clusters
-    cuml_scores = cu_silhouette_samples(X, labels, metric=metric)
+    cuml_scores = cu_silhouette_samples(X, labels, metric=metric,
+                                        chunksize=int(X.shape[0]/
+                                                      chunk_divider))
     sk_scores = sk_silhouette_samples(X, labels, metric=metric)
     assert_allclose(cuml_scores, sk_scores, rtol=1e-2, atol=1e-5)