+++ python/cuml/ensemble/randomforestregressor.pyx
@@ -712,9 +712,9 @@ class RandomForestRegressor(BaseRandomForestModel, RegressorMixin):
         del(preds_m)
         return stats
 
-    def print_summary(self):
+    def get_summary_text(self):
         """
-        Prints the summary of the forest used to train and test the model
+        Obtain the text summary of the random forest model
         """
         cdef RandomForestMetaData[float, float] *rf_forest = \
             <RandomForestMetaData[float, float]*><uintptr_t> self.rf_forest
@@ -723,14 +723,13 @@ class RandomForestRegressor(BaseRandomForestModel, RegressorMixin):
             <RandomForestMetaData[double, double]*><uintptr_t> self.rf_forest64
 
         if self.dtype == np.float64:
-            print_rf_summary(rf_forest64)
+            return get_rf_summary_text(rf_forest64).decode('utf-8')
         else:
-            print_rf_summary(rf_forest)
+            return get_rf_summary_text(rf_forest).decode('utf-8')
 
-    def print_detailed(self):
+    def get_detailed_text(self):
         """
-        Prints the detailed information about the forest used to
-        train and test the Random Forest model
+        Obtain the detailed information for the random forest model, as text
         """
         cdef RandomForestMetaData[float, float] *rf_forest = \
             <RandomForestMetaData[float, float]*><uintptr_t> self.rf_forest
@@ -739,13 +738,13 @@ class RandomForestRegressor(BaseRandomForestModel, RegressorMixin):
             <RandomForestMetaData[double, double]*><uintptr_t> self.rf_forest64
 
         if self.dtype == np.float64:
-            print_rf_detailed(rf_forest64)
+            return get_rf_detailed_text(rf_forest64).decode('utf-8')
         else:
-            print_rf_detailed(rf_forest)
+            return get_rf_detailed_text(rf_forest).decode('utf-8')
 
-    def dump_as_json(self):
+    def get_json(self):
         """
-        Dump (export) the Random Forest model as a JSON string
+        Export the Random Forest model as a JSON string
         """
         cdef RandomForestMetaData[float, float] *rf_forest = \
             <RandomForestMetaData[float, float]*><uintptr_t> self.rf_forest
@@ -754,8 +753,8 @@ class RandomForestRegressor(BaseRandomForestModel, RegressorMixin):
             <RandomForestMetaData[double, double]*><uintptr_t> self.rf_forest64
 
         if self.dtype == np.float64:
-            return dump_rf_as_json(rf_forest64).decode('utf-8')
-        return dump_rf_as_json(rf_forest).decode('utf-8')
+            return get_rf_json(rf_forest64).decode('utf-8')
+        return get_rf_json(rf_forest).decode('utf-8')
 
     def _more_tags(self):
         return {