+++ python/cuml/test/experimental/test_explainer_permutation_shap.py
@@ -27,7 +27,6 @@ from sklearn.datasets import make_classification
 from sklearn.datasets import make_regression
 from sklearn.model_selection import train_test_split
 
-
 models_config = ClassEnumerator(module=cuml)
 models = models_config.get_models()
 
@@ -47,6 +46,7 @@ def exact_tests_dataset():
     y_train = y_train.astype(np.float32)
     y_test = y_test.astype(np.float32)
     return X_train, X_test, y_train, y_test
+
 
 ###############################################################################
 #                              End to end tests                               #
@@ -73,7 +73,7 @@ def test_regression_datasets(exact_tests_dataset, model):
 
     exp_v = float(explainer.expected_value)
     fx = mod.predict(X_test)
-    assert(np.sum(cp.asnumpy(cu_shap_values)) - abs(fx - exp_v)) <= 1e-5
+    assert (np.sum(cp.asnumpy(cu_shap_values)) - abs(fx - exp_v)) <= 1e-5
 
     skmod = cuml_skl_class_dict[model]().fit(X_train, y_train)
 
@@ -84,7 +84,7 @@ def test_regression_datasets(exact_tests_dataset, model):
     skl_shap_values = explainer.shap_values(X_test)
     exp_v = float(explainer.expected_value)
     fx = mod.predict(X_test)
-    assert(np.sum(cp.asnumpy(skl_shap_values)) - abs(fx - exp_v)) <= 1e-5
+    assert (np.sum(cp.asnumpy(skl_shap_values)) - abs(fx - exp_v)) <= 1e-5
 
 
 def test_exact_classification_datasets():
@@ -112,9 +112,9 @@ def test_exact_classification_datasets():
 
     exp_v = explainer.expected_value
     fx = mod.predict_proba(X_test)[0]
-    assert(np.sum(cp.asnumpy(
+    assert (np.sum(cp.asnumpy(
         cu_shap_values[0])) - abs(fx[0] - exp_v[0])) <= 1e-5
-    assert(np.sum(cp.asnumpy(
+    assert (np.sum(cp.asnumpy(
         cu_shap_values[1])) - abs(fx[1] - exp_v[1])) <= 1e-5
 
     mod = sklearn.svm.SVC(probability=True).fit(X_train, y_train)
@@ -127,9 +127,9 @@ def test_exact_classification_datasets():
 
     exp_v = explainer.expected_value
     fx = mod.predict_proba(X_test)[0]
-    assert(np.sum(cp.asnumpy(
+    assert (np.sum(cp.asnumpy(
         skl_shap_values[0])) - abs(fx[0] - exp_v[0])) <= 1e-5
-    assert(np.sum(cp.asnumpy(
+    assert (np.sum(cp.asnumpy(
         skl_shap_values[1])) - abs(fx[1] - exp_v[1])) <= 1e-5
 
 
@@ -196,6 +196,30 @@ def test_not_shuffled_explanation(exact_tests_dataset):
 
     assert np.allclose(shap_values, not_shuffled_shap_values,
                        rtol=1e-04, atol=1e-04)
+
+
+# Test against exact shap values for linear regression
+# 1 permutation should give exact result
+def test_permutation(exact_tests_dataset):
+    X_train, X_test, y_train, y_test = exact_tests_dataset
+    # Train arbitrary model to get some coefficients
+    mod = cuml.LinearRegression().fit(X_train, y_train)
+    # Single background and foreground instance
+    # Gives zero effect to features when they are 'off'
+    # and the effect of the regression coefficient when they are 'on'
+    X_background = np.zeros((1, X_train.shape[1]))
+    X_foreground = np.ones((1, X_train.shape[1]))
+    explainer = cuml.experimental.explainer.PermutationExplainer(
+        model=mod.predict,
+        masker=X_background)
+
+    shap_values = explainer._explain(
+        X_foreground,
+        npermutations=5,
+        main_effects=False
+    )
+
+    assert np.allclose(mod.coef_, shap_values, rtol=1e-04, atol=1e-04)
 
 
 ###############################################################################