+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## New Features
 - PR #3160: Least Angle Regression (experimental)
 - PR #2659: Add initial max inner product sparse knn
+- PR #3092: Multiclass meta estimator wrappers and multiclass SVC
 - PR #2836: Refactor UMAP to accept sparse inputs
 - PR #3126: Experimental versions of GPU accelerated Kernel and Permutation SHAP
 
@@ -61,6 +62,7 @@
 - PR #3177: Make Multinomial Naive Bayes inherit from `ClassifierMixin` and use it for score
 - PR #3241: Updating RAFT to latest
 - PR #3240: Minor doc updates
+- PR #3275: Return confusion matrix as int unless float weights are used
 
 ## Bug Fixes
 - PR #3164: Expose silhouette score in Python